(chore): Refactor fnop
